+++ src/deepagents/__init__.py
@@ -1,8 +1,7 @@
 from deepagents.graph import create_deep_agent, async_create_deep_agent
 from deepagents.middleware import PlanningMiddleware, FilesystemMiddleware, SubAgentMiddleware
-from deepagents.interrupt import ToolInterruptConfig
 from deepagents.state import DeepAgentState
-from deepagents.sub_agent import SubAgent
+from deepagents.types import SubAgent, CustomSubAgent
 from deepagents.model import get_default_model
 from deepagents.builder import (
     create_configurable_agent,
