+++ src/deepagents/middleware.py
@@ -1,15 +1,19 @@
 """DeepAgents implemented as Middleware"""
 
-from langchain.agents.middleware import AgentMiddleware, AgentState, ModelRequest
-from deepagents.state import Todo, file_reducer
+from langchain.agents import create_agent
+from langchain.agents.middleware import AgentMiddleware, AgentState, ModelRequest, SummarizationMiddleware
+from langchain.agents.middleware.prompt_caching import AnthropicPromptCachingMiddleware
+from langchain_core.tools import BaseTool, tool, InjectedToolCallId
+from langchain_core.messages import ToolMessage
+from langchain.chat_models import init_chat_model
+from langgraph.types import Command
+from langgraph.prebuilt import InjectedState
 from typing import NotRequired, Annotated
+from deepagents.state import Todo, file_reducer
 from deepagents.tools import write_todos, ls, read_file, write_file, edit_file
-from deepagents.prompts import WRITE_TODOS_SYSTEM_PROMPT, TASK_SYSTEM_PROMPT, FILESYSTEM_SYSTEM_PROMPT
-from deepagents.sub_agent import SubAgent, CustomSubAgent, create_task_tool
+from deepagents.prompts import WRITE_TODOS_SYSTEM_PROMPT, TASK_SYSTEM_PROMPT, FILESYSTEM_SYSTEM_PROMPT, TASK_TOOL_DESCRIPTION, BASE_AGENT_PROMPT
+from deepagents.types import SubAgent, CustomSubAgent
 
-import logging
-
-logger = logging.getLogger(__name__)
 
 ###############################
 # Current Limitations
@@ -17,6 +21,10 @@ logger = logging.getLogger(__name__)
 # - State Schema doesn't work right now!
 # - Need to add back interrupt support!
 ###############################
+
+###########################
+# Planning Middleware
+###########################
 
 class PlanningState(AgentState):
     todos: NotRequired[list[Todo]]
@@ -35,6 +43,10 @@ class PlanningMiddleware(AgentMiddleware):
     #         "messages": [AIMessage(content="")]
     #     }
 
+###########################
+# Filesystem Middleware
+###########################
+
 class FilesystemState(AgentState):
     files: Annotated[NotRequired[dict[str, str]], file_reducer]
 
@@ -45,6 +57,10 @@ class FilesystemMiddleware(AgentMiddleware):
     def modify_model_request(self, request: ModelRequest, agent_state: AgentState) -> ModelRequest:
         request.system_prompt = request.system_prompt + "\n\n" + FILESYSTEM_SYSTEM_PROMPT
         return request
+
+###########################
+# SubAgent Middleware
+###########################
 
 class SubAgentMiddleware(AgentMiddleware):
     def __init__(
@@ -67,3 +83,129 @@ class SubAgentMiddleware(AgentMiddleware):
     def modify_model_request(self, request: ModelRequest, agent_state: AgentState) -> ModelRequest:
         request.system_prompt = request.system_prompt + "\n\n" + TASK_SYSTEM_PROMPT
         return request
+
+def _get_agents(
+    tools,
+    subagents: list[SubAgent | CustomSubAgent],
+    model
+):
+    agents = {
+        # NOTE: The general-purpose agent gets todos and files, but not subagents.
+        "general-purpose": create_agent(
+            model,
+            prompt=BASE_AGENT_PROMPT,
+            tools=tools,
+            checkpointer=False,
+            middleware=[
+                PlanningMiddleware(),
+                FilesystemMiddleware(),
+                AnthropicPromptCachingMiddleware(ttl="5m"),
+                SummarizationMiddleware(
+                    model=model,
+                    max_tokens_before_summary=20000,
+                    messages_to_keep=20,
+                ),
+            ]
+        )
+    }
+    tools_by_name = {}
+    for tool_ in tools:
+        if not isinstance(tool_, BaseTool):
+            tool_ = tool(tool_)
+        tools_by_name[tool_.name] = tool_
+    for _agent in subagents:
+        if "graph" in _agent:
+            agents[_agent["name"]] = _agent["graph"]
+            continue
+        if "tools" in _agent:
+            _tools = [tools_by_name[t] for t in _agent["tools"]]
+        else:
+            _tools = tools
+        # Resolve per-subagent model: can be instance or dict
+        if "model" in _agent:
+            agent_model = _agent["model"]
+            if isinstance(agent_model, dict):
+                # Dictionary settings - create model from config
+                sub_model = init_chat_model(**agent_model)
+            else:
+                # Model instance - use directly
+                sub_model = agent_model
+        else:
+            # Fallback to main model
+            sub_model = model
+        agents[_agent["name"]] = create_agent(
+            sub_model,
+            prompt=_agent["prompt"],
+            tools=_tools,
+            middleware=_agent["middleware"],
+            checkpointer=False,
+        )
+    return agents
+
+
+def _get_subagent_description(subagents: list[SubAgent | CustomSubAgent]):
+    return [f"- {_agent['name']}: {_agent['description']}" for _agent in subagents]
+
+
+def create_task_tool(
+    tools,
+    subagents: list[SubAgent | CustomSubAgent],
+    model,
+    is_async: bool = False,
+):
+    agents = _get_agents(
+        tools, subagents, model
+    )
+    other_agents_string = _get_subagent_description(subagents)
+
+    if is_async:
+        @tool(
+            description=TASK_TOOL_DESCRIPTION.format(other_agents=other_agents_string)
+        )
+        async def task(
+            description: str,
+            subagent_type: str,
+            state: Annotated[AgentState, InjectedState],
+            tool_call_id: Annotated[str, InjectedToolCallId],
+        ):
+            if subagent_type not in agents:
+                return f"Error: invoked agent of type {subagent_type}, the only allowed types are {[f'`{k}`' for k in agents]}"
+            sub_agent = agents[subagent_type]
+            state["messages"] = [{"role": "user", "content": description}]
+            result = await sub_agent.ainvoke(state)
+            return Command(
+                update={
+                    "files": result.get("files", {}),
+                    "messages": [
+                        ToolMessage(
+                            result["messages"][-1].content, tool_call_id=tool_call_id
+                        )
+                    ],
+                }
+            )
+    else: 
+        @tool(
+            description=TASK_TOOL_DESCRIPTION.format(other_agents=other_agents_string)
+        )
+        def task(
+            description: str,
+            subagent_type: str,
+            state: Annotated[AgentState, InjectedState],
+            tool_call_id: Annotated[str, InjectedToolCallId],
+        ):
+            if subagent_type not in agents:
+                return f"Error: invoked agent of type {subagent_type}, the only allowed types are {[f'`{k}`' for k in agents]}"
+            sub_agent = agents[subagent_type]
+            state["messages"] = [{"role": "user", "content": description}]
+            result = sub_agent.invoke(state)
+            return Command(
+                update={
+                    "files": result.get("files", {}),
+                    "messages": [
+                        ToolMessage(
+                            result["messages"][-1].content, tool_call_id=tool_call_id
+                        )
+                    ],
+                }
+            )
+    return task
